refactor(agents-desktop): extract cloud IPC handlers
Move cloud auth and cloud agent-server IPC registration into a focused registrar to keep main.ts centered on orchestration.
